Containerized the language information
This gets rid of having to update many areas if a new language came. However, it also allows many cultures to be placed in a list for the specified language.

This allows us to implement the culture selector that will be invoked when the user goes to a config entry that will allow them to specify the culture used for their current language.