Encapsulated the actual milliseconds getter
To further simplify testing the clocksource (CPU) to check to see if it's sleeping exactly the time provided.

Depending on the system (how the Spectre and Meltdown are mitigated (generic retpoline for CPUs that don't support SSE2 such as AMD Athlon XP 1500+, SSE2-LFENCE retpoline, ...), and so on), actual milliseconds sleep time may not be accurate.