Implemented custom splashes!!!
You can make use of this brand new feature to make your own custom splash, a screen that appears when Kernel Simulator boots up.

If you're bored from the Simple, Progress, or even the Plain splashes, why not make one?

However, the documentation clearing this concept is still not ready yet, but will hopefully be ready by the end of April.

The Truth: We didn't have enough time to implement this awesome feature back in Kernel Simulator 0.0.20.0, because of the internal infrastructure problems that plagued us.