Fixed race condition regarding the event/reminder polling
It looks like that we're polling for events or reminders way too fast, and if the calendar events or reminders still didn't finish loading, we're trying to reference to it, which, in turn, causes the crash because it didn't finish deserializing on time.

We've implemented synclocks and made longer delays from 1 ms to 100 ms to ensure that there will be no more crashes.