Initial color abstraction to reduce massive declarations
This is to reduce the number of times needed to declare variables needed for each new color type that will appear in future development cycles, including this one. Please note that this is not yet done, so it could result in NullReferenceExceptions and other nasty weird bugs.

Otherwise, it worked properly on the current run and the first run.