Improved paging in selection
Apparently, the displayAnswers and answers formula duo are all wrong when we started using pages to divide the answers count.

We have corrected this major error by populating how many answers per page in a different way; subtracting the window height by the current CursorTop position and 2. Why 2? because we've made two new lines as instructed in line 82 according to this commit.