Implemented new test
First off, this is going to be the new way to test kernel components.

Second off, the old test shell is going to be removed soon, which means that you'll no longer be able to provide custom parameters. However, that will change across the development cycle.

Finally, this test facility is still work in progress, so we'll start the migration process as soon as possible.