Use the new method to wait with no blocking
This finally speeds things up, because previously we have used the polling method to check to see if the thread work is alive or not. The culprit is from the Thread.Sleep(1) function where in Windows the time resolution is around 60 FPS, which means 15.7-16.0 milliseconds latency. This is problematic for screensavers in Windows, but we have successfully managed to work around this problem by finally using the Join(int) method, covered by KernelThread as Wait(int).

For more information about the latency, see https://stackoverflow.com/a/6254753

This, however, doesn't fix sleep attempts for exactly one millisecond; it only fixes the SleepNoBlock function where it repeatedly sleeps for one millisecond.

You might have to change the way you handle the thread interruptions.