[RPC-34] [BREAKING] Merging to serialization-based config
A lot of things will have to be done when performing this migration. There's an interesting backstory regarding this change.

When we're developing this version of Nitrocid KS, one of the changes that were to be implemented was that we've had to enhance the configuration reader to stop using the hard-coded values for each config entry. As a result, back on September 9th, 2022, we've implemented the new configuration reader, which was experimental at the time, found in the commit a70787eafd44c32e7ab6ad7814c587dc03d62dc9.

The configuration reader was later stabilized to be fit-for-purpose, but we've warned in the commit description that it might be slower than the earliest hard-coded-values-based configuration reader. We've promised that we would be attempting to speed this up across milestones.

The new writer, which was a reflection-based sister of the reflection-based configuration reader, first surfaced on September 20th, 2022 at commit 9d350f4b15be6bb2f01d346785800bd949e098ff, and finalized it a day later at commit 7d134f853a516defb9c494c56f2c9f0f88bd8175.

We've merged the two configuration reader/writer methods to one, effectively removing the oldest one based on hard-coding the values at commit 94d1789f05f76589d9d7eedd7b344ef67c6f12b8 released on December 23, 2022.

However, the config reader and the writer were slow because it used reflection, so we tried to use the Expressions method to attempt to speed this up, but we've only seen subtle performance improvements, as of commit ec3623f332e74a67581c204d4791624af17aff86 released at December 20th, 2022.

This commit will remove the reflection-based config reader and writer and replace them with the serialization-based config reader and writer, and we've noticed significant performance improvements regarding reading the configuration and writing them to a file.

However, as a consequence, the format is changed to satisfy the serialization conditions so that the users can easily pinpoint which part of the config is which. On the main configuration JSON file, we've removed sections to keep the serialization simple.

As a result, we've split the configuration JSON file to three files:

- KernelMainConfig.json
- KernelSaverConfig.json
- KernelSplashConfig.json

However, this implementation is rather incomplete, as the Settings app doesn't understand the changes that the new format required us to do. We've been required to make some fields to be properties, to make some internal changes to some structures, and to fix all the value errors as in the case of the "MaximumColorLevel" fields in some of the screensaver settings, which were set to zero. It also corrects it to 255 so that it can be treated as maximum.

As a result, a couple of removals had to be made. These are the functions from the configuration management module that had to be removed as of this commit:

- GetNewConfigObject()
- ReadFailsafeConfig()
- RepairConfig()
- GetConfigCategory()
- SetConfigValue()
- GetConfigValue()
- GetValueFromEntry()

We've also had to change some of the functions' signatures. Full listing is on the GitBook documentation under "Compatibility" section.

- CreateConfig()
- TryCreateConfig()
- ReadConfig()
- TryReadConfig()

Of course, there may be bugs that we haven't caught during the development, but it's mostly stable.

Good luck!