imp - doc - Make "usermanual" only show the links
Only show the links to the user manual and the developer's manual

---

When a user executes "usermanual", they're now shown two links for separate resources: user manual and the developer's manual for creating mods and screensavers. The reason for this change is that invoking Process.Start() on the URLs is Windows-exclusive, and can't be achieved in a cross-platform way.

---

Type: imp
Breaking: False
Doc Required: True
Part: 1/1