rem - brk|doc - Removed SaveCurrDir()
We have removed the SaveCurrDir() function.

---

This function used to save the current directory value to the kernel configuration file stored in the `%LOCALAPPDATA%/KS/KernelMainConfig.json` for Windows or the `$HOME/.config/ks/KernelMainConfig.json`. Following the recent improvements in the configuration system that took place two times during 0.1.0's development, the SaveCurrDir() function was effectively reduced to a single call to the Config.CreateConfig() function, which already deals with this kind of change, rendering this function useless. As a result, we decided to remove this function as it added no value and was just an extraneous wrapper function.

---

Type: rem
Breaking: True
Doc Required: True
Part: 1/1