add - brk|doc|ptp - Initial implementation of interactive TUI
An initial implementation of the interactive TUI.

---

We've made an initial implementation of the interactive TUI to make it easier to make your two-paned interactive TUI applications from your mods. This implementation is however experimental prototype and may introduce bugs in the apps that use this feature.

---

Type: add
Breaking: True
Doc Required: True
Part: 1/1