rem - brk|doc - Removed ConsoleWrapper.WindowTop
Removed a completely unnecessary function, ConsoleWrapper.WindowTop

---

First, Console.WindowTop is always 0 on both the Windows and the Unix operating systems, but you can set this value only on Windows systems, which kills the potential of cross-platform. Second, this is deemed unnecessary for most of the things, which makes it redundant. Finally, we've removed this function from both the ConsoleWrapper and the Console driver.

---

Type: rem
Breaking: True
Doc Required: True
Part: 1/1