imp - Allow selecting the last entry in selection style
We've allowed selection of the last entry in the selection style for each page.

---

Earlier, the last entry was a placeholder that said "Highlight this entry to go to the next page." Now, we've made changes so that the last selection is really selected without it being an automatic trigger to the next page.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1