add - doc - Added user custom settings
Custom settings allow different applications (built-in and mods) to put any kind of configuration to the user. It's in the format of (key, array of values).

---

We've added custom settings to allow your mod or our built-in applications, such as the RSS feed reader, to put any kind of configuration to the user. It's defined in the format of a JSON object that contain properties, which holds array of values.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1