imp - brk|doc|ptp - Initial support for SwitchInfo
We've added support for SwitchInfo

---

We've added support for SwitchInfo so that the help system can handle the switches. As a consequence, we have to phase out support for HelpUsage class, causing us to change the constructor of CommandArgumentInfo. Note that this is an initial release, so don't document anything yet about this until we finish the implementation.

---

Type: imp
Breaking: True
Doc Required: True
Part: 1/1