add - doc - Added ContainsSwitch
We've added a viable way to check to see if there is a switch by specific name

---

We've added a function, ContainsSwitch, to check to see if there is a switch by specific name. This is to alleviate the issue of using Contains against the array of switches directly.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1