add - Implemented a real Matrix
The real Matrix is here!

---

We've renamed the current implementation of Matrix to SimpleMatrix, because it doesn't resemble the real Matrix, while Matrix was overhauled to create a Matrix version of FallingLine so that we can create a real version of Matrix.

We've set this screensaver as a default.

---

Type: add
Breaking: False
Doc Required: False
Part: 1/1