add - doc - Added MkLang
We've added a sophisticated language studio tool

---

We've added MkLang, which allows you to provide translations to all the strings provided in the English strings file. Please note that it requires a working Metadata.json with appropriate language information that Nitrocid KS knows about, as well as an eng.txt file containing all strings to be translated. This language studio tool makes use of the LocaleGen components, so we had to separate the LocaleGen binary from the core library to avoid circular dependency.

We've also finalized the LocaleGen projects so that they no longer have to rely on the older AssemblyInfo making references to Kernel Simulator 0.0.20.0 released February 22nd, 2022.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1