add - doc - ShipDuet added as a game and a screensaver
We've added the ShipDuet game that supports simulation mode, so it's also added as a screensaver.

---

You can use ShipDuet to get in a fight with the two spaceships to see who wins. Two players in one computer. You can also use it a a screensaver.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1