imp - brk|doc - Time and date API is now refined
We've refined the time and date API in many ways!

---

The time and date API is now more polished with many breaking changes! We have added useful functions and improved areas of their logic so that mods can now rely on this API to use the kernel time and date more easily!

---

Type: imp
Breaking: True
Doc Required: True
Part: 1/1