fix - reg - Fixed timer and stopwatch deadlock
We've fixed timer and stopwatch not being able to exit

---

Timer and Stopwatch were unable to exit because their update threads keep running even after they're told to stop, which causes the kernel thread manager to wait for them for 60 seconds. As a result, we've changed the polling condition to (!IsStopping) so that these threads actually respond to this request when a user presses ESC during the stopwatch/timer advancement.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1