imp - Mods should work only if they're built for the sam...
...e kernel API version

We've added a strict requirement for mod finalization.

---

We've added a strict requirement for mod finalization where if the mod parser has detected that the mod API version is different from the current kernel API version, we'd fail. This is because we need to avoid errors related to TypeLoadException caused by breaking changes which are documented in our GitBook documentation website.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1