imp - Added mitigation to prevent choice-related infinit...
...e loops

We've added a mitigation strategy to prevent choice-related infinite loops.

---

When a Choice command has been provided with multiple-character answers and forgot to switch on the PressEnter variable, we now check for this and provide a helpful error message instead of an infinite loop.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1