imp - No command is specified = get list of commands in ...
...autocomplete

When no command is specified or a part of it is specified, it now queries the list of commands.

---

Instead of querying the list of available files and directories, we now need to query the list of commands so that we can automatically complete the command.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1