imp - Make the lyrics screensaver more clear about non-e...
...xistent lyrics

Lyrics screensaver should tell the user if the lyrics don't exist.

---

If the lyrics file doesn't exist, or if the path to lyrics doesn't contain any lyrics, we'll now inform the user about this instead of leaving them in the dark with no message.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1