ref - Random is no longer relevant
ScreensaverDisplayer has a field that indicates that if it's out of saver or not.

---

ScreensaverDisplayer had a field that was named OutOfRandom in an attempt to solve the problem of "Random" screensaver not working properly. However, because we've treated "random" specially, we need to rename this field to OutOfSaver so that it doesn't look like that it's only relevant for Random.

---

Type: ref
Breaking: False
Doc Required: False
Part: 1/1