imp - Fixed permission management tests false failure
We've fixed a false failure for permission management tests

---

We've accidentally placed a [Test] attribute on the [OneTimeSetUp] and the [OneTimeTearDown] functions, causing them to be run twice. This is fixed so that we can make sure that they're only run once.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1