imp - Total page count should count correctly
Total page count in the selection style should count correctly.

---

When we have AllAnswers.Count mod answersPerPage equals zero, this means that we have exactly N - 1 number of pages, so we should count the pages right. For example, if AllAnswers.Count is 16 and answersPerPage is 16, initially, pages will be 2. However, the added condition makes it 1.

Backport to Terminaux is needed.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1