fix - Fixed kernel child tests failing
Actually, there was an issue with the unit tests regarding this functionality.

---

We've finally found a fix for kernel child tests failing to stop in 5 seconds. The reason is that because there was a bug in the implementation of the functions used for testing.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1