imp - sec - Disable loading mods by default
We've disabled loading mods by default

---

When Nitrocid KS is started for the first time, loading mods are disabled by default. This means that the user will have to go to settings to enable loading mods on startup before automatically activating them on boot. This is to ensure that the default settings generated by the kernel are secure.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1