fix - Fixed NRE on arguments parser
We've fixed NRE on arguments parser

---

When the arguments parser encounters a non-existent command, it should return a valid ProvidedArgumentsInfo() with a command, its passed arguments, and a list of arguments. This way, we would have avoided NREs related to non-existent commands.

As for the ShellManager using it, it was only used for UESHParse.Execute(), taking the text of arguments. This call was removed to avoid duplicate invocations.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1