imp - Detect data source existence
We need to detect data source existence in the interactive TUI.

---

We need to detect data source to check to see if both data sources are empty or null. If they are, display a message and offer the user to exit the program.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1