add - doc - Brought back CDir!
We've added CDir from 0.0.1!

---

CDir existed back on 0.0.1 since there was no scripting, so it was removed as "useless." However, we've implemented scripting a long time ago, so we've brought it back as a command meant for scripting and for general use. We won't remove it again.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1