imp - Be more explicit about checking for 256 colors
We need to be more explicit about the 256 color checking on Windows

---

For Windows systems, instead of assuming that the console support is enabled, actually parse the console mode and warn if VT sequence processing is not enabled.

Backport to Terminaux is needed.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1