imp - Make maintenance mode work properly
We need not to say "out of chances" when rebooting from maintenance mode

---

When rebooting or shutting down the kernel from the maintenance mode, we shouldn't say "out of chances" as that is only relevant when there are three failed attempts to log-in. Also, we should actually honor the three-tries in maintenance mode.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1