imp - Better handle aliases
We've changed the behavior of GetWrappableCommands and GetCommands to get aliased commands

---

We've changed the behavior of GetWrappableCommands() and GetCommands() to get aliased commands. In consequence, we've had to remove the AliasExecutor class as it doesn't handle wrapped commands properly and is error-prone.

They have also been modified to work with mod commands, so we've had to remove ModExecutor, too. We've conducted tests with the refined system and they seemed to go well.

Both of these now-removed classes were made when the UESH shell failed to properly handle mods and aliases in the first-generation versions of Nitrocid KS. Now that we've made significant improvements regarding argument and switch parsing for all commands thanks to the new classes and functions like ArgumentsParser and ParseShellCommandArguments(), we've deicided to remove the two private classes dedicated to handle "special" cases where UESH needed help parsing both the mod and the alias commands.

This means that UESH is almost fully ready feature-wise, and this is the status that we needed to see when we were working on the first version of Nitrocid KS back on 2018, but scripting-wise, we still need to conduct a lot of work regarding this matter.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1