add - doc - Check for overflow TUI-wide
We've added LastOnOverflow() to check the selection in both panes.

---

For two panes, we've added LastOnOverflow() to check to see if the selection is after the last element. If this is the case, LastOnOverflow() will bring the cursor to the last element.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1