imp - brk|doc - Protect the Themes dictionary field
We've wrapped the Themes dictionary field with GetInstalledThemes()

---

We've made a new function called GetInstalledThemes() that allows you to get a dictionary containing a list of installed themes. Instead of having to add the theme to three places, we only have to add the theme to two places:
  - ThemeResources.resx
  - Themes/Theme.json

This is to make sure that there will be no hard-coding in ThemeTools.cs for each theme installed.

---

Type: imp
Breaking: True
Doc Required: True
Part: 1/1