add - doc - Added an option not to parse places in echo
We've added an option not to parse any placeholders in the "echo" command.

---

If you don't want to parse any placeholder in the printed text, you can pass the -noparse switch to the beginning of the "echo" command so that placeholders don't get parsed. Current behavior of parsing these placeholdes is retained for backwards-compatibility with scripts build for earlier versions of Nitrocid.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1