imp - Positional writers now handles VT sequences proper...
...ly

We've made the positional writers handle VT sequences properly.

---

To have proper handling for VT sequences, we need to let the positional writers not to increment the simulated position if it came across the VT sequence so that wrapping won't be messed up with VT sequence characters.

Backport to Terminaux is required.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1