imp - No more buffer-based workarounds
We're one workaround short!

---

Some of the legitimate writers that are found in both Nitrocid and
Terminaux, SeparatorColor, that attempt to fill the whole line with the
separator character, causing consoles to wrap the whole line to the new
row with the left position set to zero.

For writers that do use the Line argument, like the simple console plain
writer, if the left seek position is above zero after the write, the top
position will increase by one cell and the buffer check (checking
whether we're out of buffer bounds) is done.

However, filling the line, like what SeparatorColor does, requires us to
set the left seek position in GetFilteredPositions() to zero, resulting
in the top seek position to go down by one row, counting this as going
to the next row.

Earlier, when the support for Linux systems was in the experimental
stage, the separator color needed to have a Mono-specific workaround,
which did the following: (in reference to the 0.0.16.0 tag)

  - If the platform that's running Nitrocid KS is one of the *nix
    families, then it checks to see if the cursor top position is in the
    last row.
    - If it discovers that this is the case, then it moves the cursor up
      by one row.
    - If it discovers that the condition is false, nothing is done.

Between 0.0.16.0 and 0.0.24.0, we've discovered that the Mono workaround
was tweaked a bit, now using the value of OldTop, which stores the
cursor row position before the separator write.

  - In addition to the above checking logic, if that case was false and
    the second condition, which is comparing the old top value to the
    window height of the console minus three rows, is determined to be:
    - true, then it moves the cursor up by one row.
    - false, then nothing is done.

During the development of 0.1.0, Beta 2 still had the same workaround,
but, this time, was guarded by the canPosition boolean variable, which
changes depending on the state of the console:

  - If the console is dumb, this variable is false.
  - Otherwise, true.

We're glad that we've finally fixed this outstanding bug that was really
outstanding and was haunting us since 2021!

Backport to Terminaux is required.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1