imp - All shells should be killed when shutting down
We need to kill all the shells when the kernel shuts down.

---

When the kernel is shut down in any way, like rebooting or shutting down, we need to kill all the shells in order to be able to shut the kernel down in case it's not in the mother shell.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1