rem - brk|doc - Removed UserProperty
We've removed a leftover enum from UserManagement

---

Removed UserProperty, which was used while the old JSON LINQ method was used. We've migrated the user configuration to serialization and deserialization technique earlier, so we need to remove the leftover enum, which is UserProperty.

---

Type: rem
Breaking: True
Doc Required: True
Part: 1/1