imp - InferDriverTypeFromDriverInterfaceType() is testab...
...le

Added unit tests for InferDriverTypeFromDriverInterfaceType()

---

We've added unit tests for InferDriverTypeFromDriverInterfaceType(), because we made it better by consulting the known driver types list instead of relying on the if conditionals, which could suffer from mistakes, such as the repeated condition for one of the driver types. The RNG driver was excluded, making it impossible to use InferDriverTypeFromDriverInterfaceType() with RNG drivers.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1