imp - Be more clear about what keys are available
We need to be more clear about the available keys.

---

In the selection style, there are conditions that a specific key, like info and exit, is not available. In this case, just hide them instead of misleadingly showing them in the hopes that they work in invalid states.

Backport to Terminaux is required.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1