imp - Use info boxes when finishing the test
We've used informational boxes to prompt the user when the test is finished.

---

When the test is finished and the facade requires users to provide their input as to whether the test passed or failed, informational boxes are now used instead of the plain old choice-based input.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1