rem - brk|doc - PromptForSetLang() removed
An obsolete function has been removed.

---

In the language manager, PromptForSetLang() was used for prompting the user to set the language. However, it wasn't used since ChLang was removed several milestones back. ChLang was brought back much later, but without the usage of this function. As a result, we've removed this function.

---

Type: rem
Breaking: True
Doc Required: True
Part: 1/1