imp - Implemented kernel main loop
We've refactored the SignIn class from LogIn!

---

Finally, we've reached to the point where we can now easily take off the non-sign in part of SignIn() from Login class and put that main loop in its own place, MainLoop() from KernelEntry.

This was the state that we wanted to reach to ages ago!

This means that SignIn() is now a single-responsibility functions. Hopefully, there will be no more unrefactored multi-responsibility functions left.

When the login handler tries to shut the kernel down, it does this check after the whole login handler flow does its job (in case either the LoginScreen() or the UserSelector() function triggers the power management functions).

The requirement of entering the user password (if any) to shut down the simulated system is done as a security measure.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1