add - Added ranged line handle with custom colors
We've added ranged line handle with custom colors.

---

Not only did we add ranged line handle writer with support for custom colors that highlights a specific part of the file like the C compiler, gcc, but we've added support for custom colors on Nitrocid's implementation of the line handle writer.

Backport to Terminaux is required.

---

Type: add
Breaking: False
Doc Required: False
Part: 1/1