add - doc - Added basic ML.NET AI-based Assistant
We've added a very basic and incomplete ML.NET AI-based Assistant

---

The Assistant is here! We've added a very basic AI assistant using ML.NET as the backend to provide a simple help via prompts. Please note that it's only available for x64 computers.

The model is very incomplete, covering very narrow domains, so contribute your model!

Currently, it doesn't work because the build system fails to copy the necessary models zip file to the output, which is required for The Assistant to work.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1