fix - reg - Fixed settings crash trying to get/set
We've fixed a severe crash when trying to get any property value from
the settings

---

A recent commit about the custom settings feature caused the settings
app to crash, because the configuration base class was passed to the
PropertyManager as the base class instead of dynamically.

By placing the "as dynamic" suffix to the three functions, we've managed
to fix this crash.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1