rem - Removed "Path" entry from settings entries
We've removed the "Path" entry from the settings entries.

---

It was implemented when we had an intermediate level of settings reader and writer to try to improve configuration reading and writing. However, since we've moved to the serialization and deserialization techniques back on April 8th, we've decided to remove the useless key, "Path," that was only applicable to non-serialization readers and writers that were now removed from the kernel.

---

Type: rem
Breaking: False
Doc Required: False
Part: 1/1