fix - Fixed reflection tests failing
We've fixed unit test failures related to reflection

---

Specifically, part of the reflection tests were failing because we've moved the variables to be tested to their more appropriate places. As a result, we've had to change these variables to ones that are more lasting throughout the entire second development semester.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1