imp - Rely on GetWrappedSentences() to get correct posit...
...ion

We finally fixed an edge case!

---

An edge case involving text with exact length of the console window width has been fixed! We need to rely on GetWrappedSentences() to get wrapped paragraphs so that the filtered position detector gets the correct vertical position.

As a result, selection choices can now be used with the manual separator, just like we used to do in the settings app.

Backport to Terminaux is required.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1