ref - Clarify the parts of mod API version
We've clarified the purpose of the mod API version

---

NitrocidModAPIVersionMajor describes the major API version of the kernel, usually in the three-part version. The first two parts declares the API version, and the last part is a number that gets incremented by one each new release that involves breaking APIs, like 0.1.1 for 3.0.26.

NitrocidModAPIVersionChangeset describes how many breaking changes or additions are made throughout the entire development of the specific version of the kernel. It gets reset everytime there is a version that breaks APIs, like 0.1.1.

---

Type: ref
Breaking: False
Doc Required: False
Part: 1/1