imp - Better rendering of null symlinks (FTP)
When there is no info about the symlink in FTP, we should indicate so.

---

Some FTP servers, like ftp.riken.jp, may not provide information about the target file or folder that the symbolic links point to, so we need to indicate that there is no information given.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1