rem - Cleaned up localizations (1 out of 2 times)
This is meant to test the Locale Generator

---

We've cleaned up localizations from unused strings that came from earlier versions of Nitrocid KS. This commit is meant to be a test to verify that LocaleGen works as intended.

The second attempt will be done a week prior to the final release.

---

Type: rem
Breaking: False
Doc Required: False
Part: 1/1