imp - Don't stop finalizing addons if one addon failed
We need not to stop finalizing the rest of the addons if one addon failed.

---

We need not to stop finalizing the rest of the addons even if one addon has failed so that the rest of the addons get their chance to finalize. Same thing is done for addon unloading.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1