rem - Removed misleading instructions
We've removed misleading instructions in one of the settings entries.

---

One of the settings entries was written when Nitrocid KS didn't use selection style for lists. Now that we've moved to the selection style, which is easier to use anyways, we've decided to remove the misleading instrcutions from the description.

---

Type: rem
Breaking: False
Doc Required: False
Part: 1/1