add - brk|doc - Events can now hold ranges and calendars
---

We've modified the events manager so that they can handle ranged events.

Currently, the command that is used to add events doesn't support this feature, but we're working on it soon.

Please note that we've converted the config format to JSON, so this is a breaking change.

---

Type: add
Breaking: True
Doc Required: True
Part: 1/1