add - doc - Implemented initial interactive hex edit
---

We've added an initial implementation of the interactive hex editor that allows you to interactively edit a binary file without resorting to using the commands.

Please note that it's currently only a viewer, and it currently has bugs that will be solved as soon as possible.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1