add - doc - Added proper paging for interactive TUI
---

We've added proper paging support for the interactive TUI. You can now use Home and End to go to the first or the last element, while PageUp and PageDown goes to the previous or the next page.

Backport to Terminaux is required.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1