add - brk|doc - Added slash commands support for UESH shells
---

We've added support for slash command to be usable in the UESH shell. By default, when entering anything without the slashes, it shows you a tip saying that you have to provide the exact same command with the slash before it.

Also, we've added CommandText to the command parameters class to make it easier to access what command was written, allowing you to easily formulate a complete command text.

---

Type: add
Breaking: True
Doc Required: True
Part: 1/1