add - doc - PlaySound is back!
---

We've finally restored a feature that allows you to play a sound file. This feature was released on 0.0.8 (February 22nd, 2020) and cut on 0.0.11 (July 25th, 2020), because it used VLC as a backend. However, it caused software bloat, so we've decided to remove it back then.

This is the second attempt to release this feature, but this time, we've used our own library, BassBoom, instead of relying on VLC. Music file support is up to BassBoom, though, so it can only play MPEG audio files, such as MP3.

However, it currently doesn't work due to a bug in BassBoom for custom library paths, which will be resolved soon.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1