imp - brk|doc - Used Screen feature for splashes
---

We've used the brand new Screen feature for your splash screens, so this change is a breaking change as the interface change is required to take the output as a string containing a group of buffered text that would be printed to the console.

This change will make your splash screens perform faster and become resizable.

---

Type: imp
Breaking: True
Doc Required: True
Part: 1/1