imp - Music player shouldn't kill the entire kernel
---

When the BassBoom-powered music player fails to play any song for any of
the reasons, like missing audio device or audio system not working, this
error shouldn't crash the entire kernel without giving it a chance to
show an error message and to create a crash dump. It shows a message in an
info box instead.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1