add - doc - Restored built-in aliases!
---

Built-in aliases used to exist on 0.0.1 to provide you with shortcuts to long commands. Based on the 0.0.1's source code, we had the following commands that were aliased:
  - annoying-sound -> beep
  - cdir -> currentdir
  - changedir -> chdir
  - list -> ls

They were later removed on 0.0.5 as the help system went on a refactor, which allowed us to bring the shell to this point.

We're slowly working on restoring these aliases for those who want a 0.0.1 experience when executing these aliased commands.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1