fix - Fixed kernel failing to boot on Linux
---

Since the console resize listener thread was considered a critical
thread, we need to put it to the whitelist as it doesn't start on Linux
systems. This is temporary while we implement a dummy wait loop for that
thread, if it doesn't cause excessive CPU load.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1