und - prt - Converted codebase to C# (part 1)
---

This is part 1 of codebase change to C# as we're nearing the 0.1.0 Beta 3 release. 0.0.24.x now uses C# instead of Visual Studio for consistency.

This will get rid of Visual Basic completely. However, we're not done yet. This commit is more like a tag to indicate progress.

The build will fail, so patience is a virtue.

---

Type: und
Breaking: False
Doc Required: False
Part: 1/3