und - prt - Converted codebase to C# (part 3)
---

This is part 3 of codebase change to C# as we're nearing the 0.1.0 Beta 3 release. 0.0.24.x now uses C# instead of Visual Studio for consistency.

This will get rid of Visual Basic completely. However, we're not done yet. This commit is more like a tag to indicate progress.

Duplicate license headers are found in almost all C# files, so we need to find a way to remove them quickly. Consider this conversion project as done!

---

Type: und
Breaking: False
Doc Required: False
Part: 3/3