add - doc - Refactored the theme application
---

We've added a brand new function, ApplyTheme, that takes an instance of the theme information that you can get using GetThemeInfo() and sets the themes according to the assigned colors.

This is used to refactor ApplyThemeFromResources and ApplyThemeFromFile.

Also, we've added the dry theme setting to set the theme without changing the kernel settings unless either shut down or saved manually.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1