rem - prt - Obsoleted ManageAlias()
---

We've obsoleted ManageAlias() as a result of the recent evaluations regarding this function's purpose. It initially appeared on an alpha version of 0.0.5, 0.0.4.9, when we added custom alias functionality. It was served as a wrapper to both the AddAlias() and the RemoveAlias() functions and was meant to be used only in the command code.

Sadly, we have to sunset probably the oldest function that still exists as of now in the final release cycle.

We've also improved the appearance of the command as a convenience!

---

Type: rem
Breaking: False
Doc Required: False
Part: 1/2