add - Added legacy debug log style
---

We've added legacy debug log style. We would like choice over limitation, so we've decided to turn this feature back on for consistency. The modern log style was for testing, and we found it to be stable, so we've decided to bring the legacy debug log style back.

---

Type: add
Breaking: False
Doc Required: False
Part: 1/1