imp - Changed how default screensaver kernel config is s...
...et

---

The default screensaver kernel configuration has been changed so that it uses a plain property setting instead of relying on a variable in the screensaver manager. This allows important mods to have a chance to check to see if the screensaver is found.

You may have to add a screensaver checking logic in your mod code to ensure that it exists before trying to directly set it.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1