imp - brk|doc - Overhauled screensaver timeout interval
---

We've made a breaking change to the screen timeout in order to improve the experience of setting it up. It's now easier to set it up, because we've changed from milliseconds to timespan in the format of "DDD.HH:MM:SS.NNNNNN". You can use "HH:MM:SS" as the simplest form.

---

Type: imp
Breaking: True
Doc Required: True
Part: 1/1