add - doc - Added widgets
---

We've added the widgets support.

This is experimental, but we've enabled it by default as we feel that this feature is going to be one of many reasons why people will upgrade. It's experimental because Terminaux needs to be updated to implement position-width-aware writers.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1