imp - prt - Added a workaround about the Home butt...
...on

---

We've added an ugly workaround about the Home button.

This workaround fixes the About button by making it one cell wider than the Settings button if we sense that this button is not going to fill in the whole button panel.

---

Type: imp
Breaking: False
Doc Required: False
Backport Required: False
Part: 2/2