add - doc - Finalized SMultivar implementation
---

We've managed to finalize the implementation of SMultivar that used to
implement multiple variables in one settings key.

Historical info: SMultivar was first implemented in an earlier version of
Nitrocid Kernel. It was marked as obsolete before the 0.0.16.0 release. It
was later removed as of Milestone 5 of said release on April 2021 as per
commit fa27b483191ec46a65b38aa172d0e3503b8da30e.

---

Type: add
Breaking: False
Doc Required: True
Backport Required: False
Part: 1/1