ref - bkp - Get rid of KernelConfigTypes
---

This is a code refactor intended to simplify things by removing
KernelConfigTypes from ReflectionCommon. As it's an internal field, it
isn't a breaking change.

It was introduced in commit 5f15a29f602d2ac0c07087228c4a502e6f4d389f to
finalize the Settings application for serialization-based config.

We'll conduct code scanning to deprecate all properties that are wrappers
to this config (copy a value as is) and to later remove them in v0.1.3.0.

  - Backportable for 0.1.1
  - Backportable for 0.1.0

---

Type: ref
Breaking: False
Doc Required: False
Backport Required: True
Part: 1/1