imp - brk|doc - CheckConfigVariables() supports SM...
...ultivar

---

This is a breaking change that involves CheckConfigVariables(). It changes
the way it's defined so that SMultivar is supported. As a side effect,
we've removed a list of configuration names, which means that we're now
going to have to rely on kernel debugging.

---

Type: imp
Breaking: True
Doc Required: True
Backport Required: False
Part: 1/1