int - Added new servicing branch target
---

For all workflows, we've added the proposed v0.1.2.x-saas branch in case
we started development of the next version, which is 0.1.3.x.

We expect that the 0.1.3.x series is going to be the last of the 0.1.x
series before the next major series development starts.

---

Type: int
Breaking: False
Doc Required: False
Backport Required: False
Part: 1/1