add - Editable contact proof of concept
---

We've added a proof of concept to edit the contact name. This is currently
experimental and will be part of a new feature that is set for 0.1.3.

After this commit, there will be feature and localization freezes. This
means that there will no longer be new features and localizations.

This is to allow bug fixes and code finalization for 0.1.2 set to release
on February 27th.

---

Type: add
Breaking: False
Doc Required: False
Backport Required: False
Part: 1/1